requirements.txt
openai>=1.82.0
python-dotenv>=1.1.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.11.0
pydantic-settings>=2.0.0
